MAINT: Move fixtures from `test_plotting.py`
* Move fixtures from `test_plotting.py` to
`conftest.py` to make them more broadly
available to other tests.

* Change scope of fixtures in
`conftest.py` from "class" to "module"
since they can all be used across an
entire module

* Fix comments in `test_kda.py`
and `test_plotting.py`

* Remove unnecessary code from
testing fixture `results_4wl`
